implemented in function definition expression and arguments calls (ex.: $1($2)) creation in ObjectFactory of CompositeNodes fixed a buf with compare embedded fucntion because casting to int trunkated decimal point result

+++ ListFunc/src/Interpreter/EmbeddedFunctions/EmbeddedUtils.h
@@ -33,4 +33,4 @@ FunctionObject* getFunctionObject(Value& value);
 bool toBool(const Value& value);
 std::string toString(const Value& value);
 
-int compare(const Value& value1, const Value& value2);
+double compare(const Value& value1, const Value& value2);
